Add 64-bit context
Issue: APDEV-274

Depends-Change-Id: I8e496a6001c13a00c2e6170cdb9cc538ff0ad0b6
Depends-Change-Id: I56f71f1f4ad8f0e25c59f6a6e4c73a535c47e220
Depends-Change-Id: Ic8d311bf724a8726d2ff1fa27acbd0a94bfc4a30
Depends-Change-Id: I77704daf04f967936c09210f010bd59deddc54b7
Depends-Change-Id: Icc73bf55181aa1415c38d3d48b57dc4abd937304
Depends-Change-Id: Id80c269f963210785b86e1ebd05015b95379f514

+++ video/common/psb_ttm_glue.c
@@ -378,7 +378,7 @@ int psb_video_getparam(struct drm_device *dev, void *data,
 	drm_psb_msvdx_frame_info_t *current_frame = NULL;
 	uint32_t handle, i;
 	uint32_t device_info = 0;
-	uint32_t ctx_type = 0;
+	uint64_t ctx_type = 0;
 	struct psb_video_ctx *video_ctx = NULL;
 	struct msvdx_private *msvdx_priv = dev_priv->msvdx_private;
 	uint32_t imr_info[2], ci_info[2];
